% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age, in years, using birthdate and a reference date, e.g. reaction start date}
\usage{
calculate_age(birthdate, reference_date = Sys.Date())
}
\arguments{
\item{birthdate}{A vector of birthdate}

\item{reference_date}{A vector of reference date}
}
\value{
Age, in years, of an individual at a point in time (reference date).
}
\description{
Calculate age, in years, using birthdate and a reference date, e.g. reaction start date
}
