% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_all_date_cols.R
\name{mutate_all_date_cols}
\alias{mutate_all_date_cols}
\title{Mutate all character columns that have the string "date" in them}
\usage{
mutate_all_date_cols(data)
}
\arguments{
\item{data}{A data set in tibble format}
}
\value{
A new tibble where columns of \code{data} that have the string date are transformed to date fields.
}
\description{
Mutate all character columns that have the string "date" in them
}
