% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_csv.R
\name{read_all_csv}
\alias{read_all_csv}
\title{Read all csv files in a folder and create a list of tibbles}
\usage{
read_all_csv(folder, na_type = c("", "NA", -1))
}
\arguments{
\item{folder}{A folder with csv files}

\item{na_type}{A list of vectors that reflect null or NA values}
}
\value{
A list of tibbles including all csv files from a folder.
}
\description{
Read all csv files in a folder and create a list of tibbles
}
