% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_basic_ema.R
\name{theme_basic_ema}
\alias{theme_basic_ema}
\title{Change the ggplot theme to a basic ema theme}
\usage{
theme_basic_ema(font_family = "Verdana")
}
\arguments{
\item{font_family}{A font family ("Verdana" is defined but "PT Sans Narrow" works best).}
}
\value{
A new ggplot theme.
}
\description{
Change the ggplot theme to a basic ema theme
}
